Combine to CompactionCreationST
