+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/CompactionCreationST.java
@@ -17,9 +17,11 @@ package sleeper.systemtest.suite;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
 
 import sleeper.compaction.core.job.creation.strategy.impl.BasicCompactionStrategy;
 import sleeper.core.partition.PartitionTree;
+import sleeper.core.statestore.FileReference;
 import sleeper.core.util.PollWithRetries;
 import sleeper.systemtest.dsl.SleeperSystemTest;
 import sleeper.systemtest.dsl.compaction.FoundCompactionJobs;
@@ -27,20 +29,25 @@ import sleeper.systemtest.suite.testutil.Slow;
 import sleeper.systemtest.suite.testutil.SystemTest;
 
 import java.time.Duration;
+import java.util.List;
 import java.util.Map;
 import java.util.stream.LongStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.parallel.ExecutionMode.SAME_THREAD;
 import static sleeper.core.properties.table.TableProperty.COMPACTION_FILES_BATCH_SIZE;
 import static sleeper.core.properties.table.TableProperty.COMPACTION_STRATEGY_CLASS;
 import static sleeper.core.properties.table.TableProperty.TABLE_ONLINE;
+import static sleeper.core.statestore.testutils.StateStoreUpdatesWrapper.update;
 import static sleeper.systemtest.dsl.testutil.SystemTestPartitionsTestHelper.createPartitionTreeWithRecordsPerPartitionAndTotal;
 import static sleeper.systemtest.dsl.util.SystemTestSchema.DEFAULT_SCHEMA;
 import static sleeper.systemtest.suite.fixtures.SystemTestInstance.COMPACTION_CREATION;
+import static sleeper.systemtest.suite.testutil.FileReferenceSystemTestHelper.fileFactory;
 
 @SystemTest
 @Slow // Slow because it deploys its own instance so it can drain the whole compaction jobs queue
-public class CompactionCreationThroughputST {
+@Execution(SAME_THREAD)
+public class CompactionCreationST {
 
     @BeforeEach
     void setUp(SleeperSystemTest sleeper) {
@@ -73,5 +80,22 @@ public class CompactionCreationThroughputST {
         assertThat(jobs.checkFullCompactionWithPartitionsAndInputFiles(
                 partitions, "file1.parquet", "file2.parquet"))
                 .isEqualTo(FoundCompactionJobs.expectedFullCompactionCheckWithJobs(65536));
+    }
+
+    @Test
+    void shouldSendBadCompactionBatchToDeadLetterQueue(SleeperSystemTest sleeper) throws Exception {
+        // Given
+        sleeper.tables().createWithProperties("test", DEFAULT_SCHEMA, Map.of(TABLE_ONLINE, "false"));
+        FileReference inputFile = fileFactory(sleeper).rootFile("input.parquet", 100);
+        sleeper.stateStore().fakeCommits().setupStateStore(store -> {
+            update(store).addFile(inputFile);
+            update(store).assignJobId("job-1", List.of(inputFile));
+        });
+
+        // When
+        sleeper.compaction()
+                .sendFakeSingleCompactionBatch("job-2", List.of(inputFile))
+                // Then
+                .waitForCompactionBatchOnDeadLetterQueue();
     }
 }
